refactor(menu): extract shared navigation building blocks
