+++ apps/frontend/src/components/menu/PublicMenu.tsx
@@ -1,17 +1,16 @@
 'use client';
-import { CollapseMenuButton } from '@/components/menu/CollapseMenuButton';
+import { MenuFooter } from '@/components/menu/MenuFooter';
 import { MenuLogo } from '@/components/menu/MenuLogo';
 import PublicNavigation from '@/components/menu/PublicNavigation';
 import { cn } from '@/lib/utils';
-import { LogoFiligranIcon } from '@filigran/icon';
-import { useTranslations } from 'next-intl';
+import { useLocale } from 'next-intl';
 import { useCallback } from 'react';
 import { useLocalStorage } from 'usehooks-ts';
 
 const PublicMenu = () => {
+  const locale = useLocale();
   const [open, setOpen] = useLocalStorage<boolean>('is-public-menu-open', true);
   const handleOpenMenu = useCallback(() => setOpen((prev) => !prev), [setOpen]);
-  const t = useTranslations();
 
   return (
     <aside
@@ -19,28 +18,16 @@ const PublicMenu = () => {
         'max-md:hidden z-20 flex h-full flex-col overflow-x-hidden bg-page-background text-muted-foreground duration-300',
         open ? 'w-45' : 'w-14'
       )}>
-      <MenuLogo withDarkBackground={false} />
+      <MenuLogo
+        href={`/${locale}`}
+        withDarkBackground={false}
+      />
       <div className="flex flex-col flex-1 justify-between min-h-0">
         <PublicNavigation open={open} />
-        <div>
-          <CollapseMenuButton
-            open={open}
-            handleOpenMenu={handleOpenMenu}
-            isHomepageV2={true}
-          />
-          {open ? (
-            <div className="flex items-center px-m pb-s gap-1 text-muted-foreground/50 text-[10px] whitespace-nowrap">
-              {t('App.MadeBy')}
-              <LogoFiligranIcon className="size-3 shrink-0" />
-              {/* eslint-disable-next-line xtm-hub-i18n-rules/no-literal-string-in-jsx */}
-              {'Filigran'}
-            </div>
-          ) : (
-            <div className="flex justify-center pb-s text-muted-foreground/50">
-              <LogoFiligranIcon className="size-3" />
-            </div>
-          )}
-        </div>
+        <MenuFooter
+          open={open}
+          handleOpenMenu={handleOpenMenu}
+        />
       </div>
     </aside>
   );
