+++ apps/frontend/src/components/menu/SharedNavigation.tsx
@@ -1,0 +1,79 @@
+import { NavigationLinkMenu } from '@/components/menu/NavigationLinks';
+import {
+  ClosedSection,
+  LinkedSection,
+  OpenedSection,
+} from '@/components/menu/NavigationSections';
+import {
+  BottomLink,
+  SectionConfig,
+} from '@/components/menu/use-navigation-type';
+import { Accordion, Separator } from '@filigran/ui';
+
+interface SharedNavigationProps {
+  open: boolean;
+  sections: SectionConfig[];
+  bottomLinks: BottomLink[];
+}
+
+export const SharedNavigation = ({
+  open,
+  sections,
+  bottomLinks,
+}: SharedNavigationProps) => {
+  return (
+    <nav className="flex-1 min-h-0 overflow-y-auto">
+      {open ? (
+        <Accordion
+          type="single"
+          collapsible
+          className="w-full">
+          {sections.map((section) =>
+            section.href ? (
+              <LinkedSection
+                key={section.key}
+                section={section}
+                open={true}
+              />
+            ) : (
+              <OpenedSection
+                key={section.key}
+                section={section}
+              />
+            )
+          )}
+        </Accordion>
+      ) : (
+        <ul>
+          {sections.map((section) => (
+            <li key={section.key}>
+              {section.href ? (
+                <LinkedSection
+                  section={section}
+                  open={false}
+                />
+              ) : (
+                <ClosedSection section={section} />
+              )}
+            </li>
+          ))}
+        </ul>
+      )}
+
+      <Separator className="my-s" />
+      <ul className="space-y-s pb-s">
+        {bottomLinks.map((link) => (
+          <li key={link.key}>
+            <NavigationLinkMenu
+              open={open}
+              href={link.href}
+              icon={link.icon}
+              text={link.label}
+              external={link.external}
+            />
+          </li>
+        ))}
+      </ul>
+    </nav>
+  );
+};
