+++ apps/frontend/src/components/menu/navigation-styles.ts
@@ -1,0 +1,6 @@
+export const NAVIGATION_ACTIVE_CLASSES =
+  'bg-primary/10 shadow-[inset_2px_0px] shadow-primary text-primary';
+export const NAVIGATION_HOVER_CLASSES =
+  'hover:shadow-[inset_2px_0px] hover:shadow-white';
+export const NAVIGATION_GRADIENT_STYLE =
+  'linear-gradient(to right, hsl(var(--blue-default)), hsl(var(--turquoise-300)))';
